cherry pick theme restructuring
