+++ packages/react-theme/src/themes/index.ts
@@ -1,0 +1,2 @@
+export * from './teams';
+export * from './web';
