+++ packages/react-theme/src/themes/teams/defaultTheme.ts
@@ -1,0 +1,25 @@
+import { createNeutralColorTokens, sharedColorTokens } from '../default';
+import { createShadowLevelTokens } from '../default';
+import { brandColors, sharedColors, fontFamilies, fontWeights, fontSizes, lineHeights } from '../global';
+import { Theme } from '../types';
+
+const common = {
+  sharedColors,
+  sharedColorTokens,
+  fontFamilies,
+  fontWeights,
+  fontSizes,
+  lineHeights,
+};
+
+const teamsNeutralColorTokens = createNeutralColorTokens(brandColors.teams);
+
+export const teamsDefaultTheme: Theme = {
+  brandColors: brandColors.teams,
+  neutralColorTokens: teamsNeutralColorTokens,
+  shadowLevels: createShadowLevelTokens(
+    teamsNeutralColorTokens.neutralShadowAmbient,
+    teamsNeutralColorTokens.neutralShadowKey,
+  ),
+  ...common,
+};
