+++ packages/react-theme/src/themes/teams/index.ts
@@ -1,0 +1,3 @@
+export * from './defaultTheme';
+export * from './darkTheme';
+export * from './highContrastTheme';
