+++ packages/react-theme/src/themes/web/darkTheme.ts
@@ -1,0 +1,25 @@
+import { createNeutralColorTokens, sharedColorTokens } from '../../dark';
+import { createShadowLevelTokens } from '../../default';
+import { brandColors, sharedColors, fontFamilies, fontWeights, fontSizes, lineHeights } from '../../global';
+import { Theme } from '../../types';
+
+const common = {
+  sharedColors,
+  sharedColorTokens,
+  fontFamilies,
+  fontWeights,
+  fontSizes,
+  lineHeights,
+};
+
+const webNeutralColorTokens = createNeutralColorTokens(brandColors.web);
+
+export const webDarkTheme: Theme = {
+  brandColors: brandColors.web,
+  neutralColorTokens: webNeutralColorTokens,
+  shadowLevels: createShadowLevelTokens(
+    webNeutralColorTokens.neutralShadowAmbientDarker,
+    webNeutralColorTokens.neutralShadowKeyDarker,
+  ),
+  ...common,
+};
