+++ packages/react-theme/src/themes/web/defaultTheme.ts
@@ -1,0 +1,25 @@
+import { createNeutralColorTokens, sharedColorTokens } from '../../default';
+import { createShadowLevelTokens } from '../../default';
+import { brandColors, sharedColors, fontFamilies, fontWeights, fontSizes, lineHeights } from '../../global';
+import { Theme } from '../../types';
+
+const common = {
+  sharedColors,
+  sharedColorTokens,
+  fontFamilies,
+  fontWeights,
+  fontSizes,
+  lineHeights,
+};
+
+const webNeutralColorTokens = createNeutralColorTokens(brandColors.web);
+
+export const webDefaultTheme: Theme = {
+  brandColors: brandColors.web,
+  neutralColorTokens: webNeutralColorTokens,
+  shadowLevels: createShadowLevelTokens(
+    webNeutralColorTokens.neutralShadowAmbient,
+    webNeutralColorTokens.neutralShadowKey,
+  ),
+  ...common,
+};
