+++ packages/react-theme/src/themes/web/highContrastTheme.ts
@@ -1,0 +1,17 @@
+import { createNeutralColorTokens, sharedColorTokens } from '../../highContrast';
+import { brandColors, sharedColors, fontFamilies, fontWeights, fontSizes, lineHeights } from '../../global';
+import { Theme } from '../../types';
+const common = {
+  sharedColors,
+  sharedColorTokens,
+  fontFamilies,
+  fontWeights,
+  fontSizes,
+  lineHeights,
+};
+
+export const webHighContrastTheme: Theme = {
+  brandColors: brandColors.web,
+  neutralColorTokens: createNeutralColorTokens(brandColors.web),
+  ...common,
+};
